fix(security): slice memory allocation with excessive size value.

+++ authnode/server.go
@@ -109,8 +109,8 @@ func (m *Server) checkConfig(cfg *config.Config) (err error) {
 	if m.id, err = strconv.ParseUint(cfg.GetString(ID), 10, 64); err != nil {
 		return fmt.Errorf("%v,err:%v", proto.ErrInvalidCfg, err.Error())
 	}
-	m.config.heartbeatPort = cfg.GetInt64(heartbeatPortKey)
-	m.config.replicaPort = cfg.GetInt64(replicaPortKey)
+	m.config.heartbeatPort = cfg.GetInt(heartbeatPortKey)
+	m.config.replicaPort = cfg.GetInt(replicaPortKey)
 	if m.config.heartbeatPort <= 1024 {
 		m.config.heartbeatPort = raftstore.DefaultHeartbeatPort
 	}
@@ -162,8 +162,8 @@ func (m *Server) createRaftServer(cfg *config.Config) (err error) {
 		NodeID:            m.id,
 		RaftPath:          m.walDir,
 		NumOfLogsToRetain: m.retainLogs,
-		HeartbeatPort:     int(m.config.heartbeatPort),
-		ReplicaPort:       int(m.config.replicaPort),
+		HeartbeatPort:     m.config.heartbeatPort,
+		ReplicaPort:       m.config.replicaPort,
 		TickInterval:      m.tickInterval,
 		ElectionTick:      m.electionTick,
 	}
